update: rename src/bundler
