+++ src/model/specification.ts
