+++ src/model/stub-specification.ts
